Refactored Out Use of Cache to Store Results
Due to how fast wasm is to find the solutions for a day / month combo,
it was decided that the use of a cache to store previously found
solutions would be wasteful of memory. The code was refactored to remove
the use of a map. Now click button will always call solve and return a
solution.