Implemented Displaying Solutions
Implemented the required code to add the number of pieces to the board
based on the number of hints / pieces the user wants
shown. The colour and the order of the pieces is shuffled to
prevent the same colour and piece being shown.
The current solution will be cleared if
the user selects a day and month which has not been solved for, and the
solution will be shown if the user re-selects the day and month the
solution is solved for.

Current bugs: If the solve button is clicked multiple times, dispite
being disabled, the solve_click function will be called as soon as the
current function completes processing. This will only happen twice,
despite the number of times clicked while disabled.