Hotfix Fixed Bug with Solve button
Previously, when the solve button was pressed, it could be pressed again
while executing and immediately que up a function call. Now, a guard
statement in combination with a delayed setting of a flag is used to
prevent the solve_click function from queing when the button is pressed
while it is solving.

Also corrected some Rust imports which were wrong due to the crate being
renamed to from dfsolver to wasm-dfsolver.