Created second layout to apply to game
Previously, Svelte was applying the default layout settings to all
routes. The default layout was not appropriate for the drag-n-drop game
and a new layout file was created to be applied specifically to the
drag-n-drop game. This layout fixes the canvas from growing/shrinking
and keeps it at the dimensions specified.