Created Loading Page and refactored solver
This commit refactors the solver as it was identified that the WASM
assest was downloading every time the user clicked solve. This assest
only needs to be downloaded once, and the code was refactored to make
this happen. A simple loading spinner was also added to allow the
webpage to download the required content before it is ready for user
interation.

Bug-fix: Previously, when a new date-month is selected, the previously
displayed solution was removed. However, when either the number of hints
or solution number was changed, the pieces would be shown again. This
commit fixes this behaviour so that the pieces do not show when these
variables are changed by the user.