Implemts rotate and resset functions for pieces
Implements the rotate and reset functions for the pieces, when the
relevant buttons are pressed when a piece is selected. Also added
tweened values for theta (angle) and position when not being dragged.

TODO: Implement Flip function, and add event listeners for key presses
as well (wasd and arrow keys and space bar).