Fully translates piece in rust source code
Previously, each intermediate translation was tested as a legitimate
unique piece. This is not the case and now the pieces are fully
translated to the left until the first piece non-hole square aligns with
the boards first available square. The wasm binary files are also being
added to the remote repo.