Updates css styles
Updates calendar css so that text is centred for IOS devices. Also sets
the text for the solver button to black. Shrunk katex text which
overflowed container and caused issues for smaller devices. Adds
additional media querys to adjust padding and margins for footers and
main div. Build files updated.